# Speed up thermo calculations