--- a/src/engine/gasfun.jl
+++ b/src/engine/gasfun.jl
@@ -1,5 +1,7 @@
 using Zygote
 
+include("gasdata.jl")
+
 """
     gasfun(igas, t)